requirements.txt
speechrecognition
lyricsgenius
python-dotenv